Workathon/certiV4/requirements.txt
Flask
pandas
firebase-admin
reportlab
python-dotenv
qrcode